+++ test/dartino_util_test.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library atom.dependencies_test;
+library atom.dartino_util_test;
 
 import 'package:atom_dartlang/dartino/dartino_util.dart';
 import 'package:test/test.dart';
